XML part 2
In this commit was trying to use the XML handling from 25th tutorial in
irrlicht. But because our application starts without Irrlicht having
it's separate base I can't initialise the XML reader in a class,
therefore to do this I will need to put Irrlicht in separate base class
and then derive some values for the XML reader to work properly. This
will also make the code more readable and most of the functionality will
be hidden from the user to avoid confusion. Don't hate me guys ;)